Removed JSON dependency in instantiation model class
